Challenge_Object_Detection/requirements.txt
opencv-python==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
tqdm==4.66.2
torch==2.2.0+cu118 --index-url https://download.pytorch.org/whl/cu118
torchvision==0.17.0+cu118 --index-url https://download.pytorch.org/whl/cu118
torchaudio==2.2.0+cu118 --index-url https://download.pytorch.org/whl/cu118
ultralytics==8.1.34
pycocotools==2.0.7
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2